Add Create, Update, Like & Unlike API in Photo Module

+++ app/Http/Controllers/API/PhotoController.php
@@ -3,15 +3,24 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\ApiController;
+use App\Http\Controllers\Traits\PhotoTrait;
 use App\Http\Request\FormPhotoRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
+use App\Enum\Message;
+use App\Helper\Helper;
 
-use App\Http\Controllers\Traits\PhotoTrait;
+use DB;
 
 class PhotoController extends ApiController
 {
     use PhotoTrait;
+
+    public function __construct()
+    {
+        $this->helper = new Helper;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -46,9 +55,51 @@ class PhotoController extends ApiController
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(FormPhotoRequest $request)
     {
-        //
+        try {
+            // Make sure the user is logged in
+            if(!Auth::check()) {
+                return $this->sendResponse(401, Message::UNAUTHORIZED);
+            }
+            $user = Auth::user();
+
+            if(!$request->tags) return $this->sendResponse(400, Message::TAG_REQUIRED);
+            
+            if($request->hasFile('image')) {
+                DB::beginTransaction();
+                $image = $request->file('image');
+                $filePath = "user-$user->id";
+                // Save Image in Storage
+                $uploadStorage = $this->uploadStorage($image,$filePath);
+
+                // Set Data
+                $photo = new \StdClass;
+                $photo->user_id = $user->id;
+                $photo->title = $request->title;
+                $photo->caption = $request->caption;
+                $photo->file_name = $uploadStorage['fileName'];
+                $photo->file_path = $filePath;
+                $photo->tags = $request->tags;
+
+                $this->createPhoto($user,$photo);
+
+                $response = [
+                    'code' => 200,
+                    'message' => 'Upload photo success.'
+                ];
+                DB::commit();
+            }else {
+                $response = [
+                    'code' => 400,
+                    'message' => Message::NOT_IMAGE
+                ];
+            }
+            return $this->sendResponse($response['code'], $response['message']);
+        } catch(Exception $e) {
+            DB::rollback();
+            return $this->sendException($e);
+        }
     }
 
     /**
@@ -70,7 +121,7 @@ class PhotoController extends ApiController
             }else {
                 $response = [
                     'code' => 404,
-                    'message' => 'Data not found.'
+                    'message' => Message::NOT_FOUND
                 ];
             }
             return $this->sendResponse($response['code'], $response['message'], $photo);
@@ -105,22 +156,50 @@ class PhotoController extends ApiController
             // Find Photo by Id
             $photo = $this->findPhoto($id);
 
-            if($photo) {
-                $this->editPhoto($photo,$request);
+            // Make sure the user is logged in
+            if(!Auth::check()) {
+                return $this->sendResponse(401, Message::UNAUTHORIZED);
+            }
+            $user = Auth::user();
+
+            if($photo && $photo->user_id === $user->id) {
+                DB::beginTransaction();
+                if($request->hasFile('image')) {
+                    $image = $request->file('image');
+                    $filePath = "user-$user->id";
+                    $hasFile = true;
+                    // Save Image in Storage
+                    $uploadStorage = $this->uploadStorage($image,$filePath);
+                }else {
+                    $hasFile = false;
+                }
+                
+                // Set Data
+                $form = new \StdClass;
+                $form->user_id = $user->id;
+                $form->title = $request->title;
+                $form->caption = $request->caption;
+                $form->file_name = ($hasFile == true) ? $uploadStorage['fileName'] : $photo->file_name;
+                $form->file_path = ($hasFile == true) ? $filePath : $photo->file_path;
+                $form->tags = $request->tags;
+                
+                $this->editPhoto($user,$photo,$form);
 
                 $response = [
                     'code' => 200,
                     'message' => 'Update photo success.'
                 ];
+                DB::commit();
             }else {
                 $response = [
                     'code' => 404,
-                    'message' => 'Data not found.'
+                    'message' => Message::NOT_FOUND
                 ];
             }
             
             return $this->sendResponse($response['code'], $response['message']);
         } catch(Exception $e) {
+            DB::rollback();
             return $this->sendException($e);
         }
     }
@@ -136,13 +215,35 @@ class PhotoController extends ApiController
         try {
             // Find Photo by ID
             $photo = $this->findPhoto($id);
+
+            // Make sure the user is logged in
+            if(!Auth::check()) {
+                return $this->sendResponse(401, Message::UNAUTHORIZED);
+            }
+            $user = Auth::user();
+
             // Delete Photo by ID
             if($photo) {
-                $photo->delete();
-                return $this->sendResponse(200, 'Delete photo success.');
+                if($photo->user_id === $user->id) {
+                    $photo->delete();
+
+                    $response = [
+                        'code' => 200,
+                        'message' => 'Delete photo success.'
+                    ];
+                }else {
+                    $response = [
+                        'code' => 403,
+                        'message' => Message::FORBIDDEN
+                    ];
+                }
             }else {
-                return $this->sendResponse(404, 'Data not found.');
+                $response = [
+                    'code' => 404,
+                    'message' => Message::NOT_FOUND
+                ];
             }
+            return $this->sendResponse($response['code'], $response['message']);
         } catch(Exception $e) {
             return $this->sendException($e);
         }
@@ -157,7 +258,38 @@ class PhotoController extends ApiController
      */
     public function like(Request $request, $id)
     {
+        try {
+            // Find Photo by ID
+            $photo = $this->findPhoto($id);
 
+            // Make sure the user is logged in
+            if(!Auth::check()) {
+                return $this->sendResponse(401, Message::UNAUTHORIZED);
+            }
+            $user = Auth::user();
+            
+            if($photo) {
+                // Check if the user has liked before
+                if(!$user->likes()->where('photo_id', $photo->id)->exists()) {
+                    $photo->likes()->create([
+                        'user_id' => $user->id
+                    ]);
+                }
+
+                $response = [
+                    'code' => 200,
+                    'message' => 'Photo liked'
+                ];
+            }else {
+                $response = [
+                    'code' => 400,
+                    'message' => 'Failed to like the photo'
+                ];
+            }
+            return $this->sendResponse($response['code'], $response['message']);
+        } catch(Exception $e) {
+            return $this->sendException($e);
+        }
     }
 
     /**
@@ -169,6 +301,36 @@ class PhotoController extends ApiController
      */
     public function unlike(Request $request, $id)
     {
+        try {
+            // Find Photo by ID
+            $photo = $this->findPhoto($id);
 
+            // Make sure the user is logged in
+            if(!Auth::check()) {
+                return $this->sendResponse(401, Message::UNAUTHORIZED);
+            }
+            $user = Auth::user();
+            
+            if($photo) {
+                // Check if the user has liked before
+                if($user->likes()->where('photo_id', $photo->id)->exists()) {
+                    // Delete liked photo
+                    $photo->likes()->delete();
+                }
+
+                $response = [
+                    'code' => 200,
+                    'message' => 'Photo unliked'
+                ];
+            }else {
+                $response = [
+                    'code' => 400,
+                    'message' => 'Failed to ulike the photo'
+                ];
+            }
+            return $this->sendResponse($response['code'], $response['message']);
+        } catch(Exception $e) {
+            return $this->sendException($e);
+        }
     }
 }
